Upgrade SO for weapons
Added upgrade functionality for all the weapons.
Found bug: Need to cap the upper bound of the upgrades so when max is reached it will not let you buy more. (check level of focus weapon to see if button needs deactivating.)
Found bug: focus UI is allowing for multiple to be selected