Many Changes
GitHub App wasnt working. Here are a lot of changes.
The big one is a heystore has been made and the app has been added to the google play console. Which means we can setup OAuth and cloud saving soon.
Added the wave system and switch to android release.
Made the change to URP and updated textures.